chore: move data to app directory
