+++ src/app/_data/emotes.ts
