+++ src/app/arena/arena.component.spec.ts
@@ -1,6 +1,6 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { hallOfFame } from 'src/assets/data/hallOfFame';
+import { hallOfFame } from 'src/app/_data/hallOfFame';
 
 import { ArenaComponent } from './arena.component';
