+++ src/app/library/library.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { library } from 'src/assets/data/library';
+import { library } from 'src/app/_data/library';
 
 import { LibraryComponent } from './library.component';
