+++ src/app/museum/museum.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { emotes } from 'src/assets/data/emotes';
-import { portraits } from 'src/assets/data/portraits';
+import { emotes } from 'src/app/_data/emotes';
+import { portraits } from 'src/app/_data/portraits';
 import { Emote } from 'src/interfaces/Emote';
 import { Portrait } from 'src/interfaces/Portrait';
